Separate storage helpers based on being internal or external
